XXX: move the sink/source/const/call funcs into their own package
No need to have those in simple/ which is a simple API for new
functions.
