+++ lang/funcs/structs/call.go
@@ -15,8 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-// package structs
-package simple
+package structs
 
 import (
 	"context"
